reorganize for use with flit and pypi
to install for development: flit install -s
to build and publish: flit publish
