+++ dev-requirements.in
@@ -1,0 +1,1 @@
+flit
