+++ dev-requirements.txt
@@ -1,0 +1,9 @@
+certifi==2019.9.11        # via requests
+chardet==3.0.4            # via requests
+docutils==0.15.2          # via flit
+flit-core==2.0.2          # via flit
+flit==2.0.2
+idna==2.8                 # via requests
+pytoml==0.1.21            # via flit, flit-core
+requests==2.22.0          # via flit
+urllib3==1.25.7           # via requests
