--- install
@@ -1,3 +1,0 @@
-#!/usr/bin/bash
-pip3 install -r requirements.txt
-echo -en "\n\n\x1b[6;30;47m[+]Append \x1b[6;30;41m$PWD\x1b[6;30;47m , to your system PATH\n[+]Now on you can invoke $ tgnize path-to-exported-chat-dir path-to-sink-dir\n\x1b[0m"
