--- model/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/python3
-
-from sys import path
-from os.path import dirname
-path.append(dirname(__file__))
