+++ requirements.in
@@ -1,0 +1,4 @@
+matplotlib
+typing
+beautifulsoup4
+lxml
