--- tgnize
@@ -1,2 +1,0 @@
-#!/usr/bin/bash
-python3 `dirname $0`/app.py $@
