+++ tgnize/__init__.py
@@ -1,0 +1,3 @@
+"""How about another exported Telegram Chat analyzer? 😉"""
+
+__version__ = '0.1.2'
